Improve coder tests (#4664)
* Refactor coder tests

* Fixes

* fix scala 2.12 errors

* One more fix

* One more fix

* fix

* one more fix

* one more fix

* one more fix

* added file headers

* added more checks on primitives

* Fixing checks failures

* Fixing checks

* Fix tests

* Removed coderShouldThrowOn

* fix

* Added more checks and fixed a few coders

* Update scio-test/src/test/scala/com/spotify/scio/coders/CoderTest.scala


* fixes after review

* Fixing WrappedArray issue

* fix format issue

* Ran scalafixAll on scio-core

* modify syntax with custom options

* fix

---------
