+++ scio-core/src/main/scala/com/spotify/scio/coders/LowPriorityCoderDerivation.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Spotify AB.
+ * Copyright 2023 Spotify AB.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
