+++ scio-core/src/main/scala/com/spotify/scio/coders/instances/JodaCoders.scala
@@ -49,7 +49,7 @@ object JodaCoders {
     }
 }
 
-final private class JodaDateTimeCoder extends AtomicCoder[DateTime] {
+final private[coders] class JodaDateTimeCoder extends AtomicCoder[DateTime] {
   override def encode(value: DateTime, os: OutputStream): Unit = {
     val dos = new DataOutputStream(os)
 
@@ -65,9 +65,11 @@ final private class JodaDateTimeCoder extends AtomicCoder[DateTime] {
 
     new DateTime(ms, zone)
   }
+
+  override def consistentWithEquals() = true
 }
 
-final private class JodaLocalDateTimeCoder extends AtomicCoder[LocalDateTime] {
+final private[coders] class JodaLocalDateTimeCoder extends AtomicCoder[LocalDateTime] {
   import JodaCoders.checkChronology
 
   override def encode(value: LocalDateTime, os: OutputStream): Unit = {
@@ -96,9 +98,11 @@ final private class JodaLocalDateTimeCoder extends AtomicCoder[LocalDateTime] {
 
     new LocalDateTime(year, month, day, hour, minute, second, ms)
   }
+
+  override def consistentWithEquals() = true
 }
 
-final private class JodaLocalDateCoder extends AtomicCoder[LocalDate] {
+final private[coders] class JodaLocalDateCoder extends AtomicCoder[LocalDate] {
   import JodaCoders.checkChronology
 
   override def encode(value: LocalDate, os: OutputStream): Unit = {
@@ -119,9 +123,11 @@ final private class JodaLocalDateCoder extends AtomicCoder[LocalDate] {
 
     new LocalDate(year, month, day)
   }
+
+  override def consistentWithEquals() = true
 }
 
-final private class JodaLocalTimeCoder extends AtomicCoder[LocalTime] {
+final private[coders] class JodaLocalTimeCoder extends AtomicCoder[LocalTime] {
   import JodaCoders.checkChronology
 
   override def encode(value: LocalTime, os: OutputStream): Unit = {
@@ -144,4 +150,6 @@ final private class JodaLocalTimeCoder extends AtomicCoder[LocalTime] {
 
     new LocalTime(hour, minute, second, ms)
   }
+
+  override def consistentWithEquals() = true
 }
