+++ scio-test/src/main/scala/com/spotify/scio/testing/CoderAssertions.scala
@@ -17,9 +17,11 @@
 
 package com.spotify.scio.testing
 
-import com.spotify.scio.coders.{Coder, CoderMaterializer}
+import com.spotify.scio.coders._
+import org.apache.beam.sdk.coders.Coder.NonDeterministicException
 import org.apache.beam.sdk.coders.{Coder => BCoder}
 import org.apache.beam.sdk.options.{PipelineOptions, PipelineOptionsFactory}
+import org.apache.beam.sdk.testing.CoderProperties
 import org.apache.beam.sdk.util.{CoderUtils, SerializableUtils}
 import org.scalactic.Equality
 import org.scalatest.Assertion
@@ -30,102 +32,188 @@ import scala.reflect.ClassTag
 object CoderAssertions {
   private lazy val DefaultPipelineOptions = PipelineOptionsFactory.create()
 
-  implicit class CoderAssertionsImplicits[T](private val value: T) extends AnyVal {
+  type CoderAssertion[T] = AssertionContext[T] => Assertion
+  type CoderAssertionBase = AssertionContextBase => Assertion
+
+  case class WithOptions(opts: PipelineOptions)
+
+  trait CustomOptionsSyntax[T] {
+    def should(coderAssertion: CoderAssertion[T]): AssertionContext[T]
+  }
+
+  implicit class ValueShouldSyntax[T](value: T) {
     def coderShould(
       coderAssertion: CoderAssertion[T]
-    )(implicit c: Coder[T], eq: Equality[T]): Assertion =
-      coderAssertion.assert(value)
+    )(implicit c: Coder[T]): AssertionContext[T] = {
+      val ctx = AssertionContext(Some(value), c)
+      ctx.copy(lastAssertion = Some(coderAssertion(ctx)))
+    }
+
+    def kryoCoderShould(
+      coderAssertion: CoderAssertion[T]
+    )(implicit ct: ClassTag[T]): AssertionContext[T] = {
+      val ctx = AssertionContext(Some(value), Coder.kryo[T])
+      ctx.copy(lastAssertion = Some(coderAssertion(ctx)))
+    }
+
+    def coder(
+      optionsTerm: WithOptions
+    )(implicit c: Coder[T]): CustomOptionsSyntax[T] = new CustomOptionsSyntax[T] {
+      override def should(coderAssertion: CoderAssertion[T]): AssertionContext[T] = {
+        val ctx = AssertionContext(Some(value), c, opts = optionsTerm.opts)
+        ctx.copy(lastAssertion = Some(coderAssertion(ctx)))
+      }
+    }
   }
 
-  trait CoderAssertion[T] {
-    def assert(value: T)(implicit c: Coder[T], eq: Equality[T]): Assertion
+  implicit class CoderShouldSyntax[T](c: Coder[T]) {
+    def coderShould(
+      coderAssertion: CoderAssertion[T]
+    ): AssertionContext[T] = {
+      val ctx = AssertionContext(None, c)
+      ctx.copy(lastAssertion = Some(coderAssertion(ctx)))
+    }
   }
 
-  def roundtripWithCustomAssert[T](
+  case class AssertionContext[T](
+    actualValue: Option[T],
+    coder: Coder[T],
+    lastAssertion: Option[Assertion] = None,
     opts: PipelineOptions = DefaultPipelineOptions
-  )(customAssertEquality: (T, T) => Assertion): CoderAssertion[T] =
-    new CoderAssertion[T] {
-      override def assert(value: T)(implicit c: Coder[T], eq: Equality[T]): Assertion = {
-        val beamCoder = CoderMaterializer.beamWithDefault(c, o = opts)
-        val result = roundtripWithCoder(beamCoder, value)
-        customAssertEquality(value, result)
-      }
-    }
+  ) extends AssertionContextBase {
+    override type ValType = T
 
-  def roundtrip[T](
-    opts: PipelineOptions = DefaultPipelineOptions
-  ): CoderAssertion[T] =
-    new CoderAssertion[T] {
-      override def assert(value: T)(implicit c: Coder[T], eq: Equality[T]): Assertion = {
-        val beamCoder = CoderMaterializer.beamWithDefault(c, o = opts)
-        checkRoundtripWithCoder(beamCoder, value)
-      }
-    }
+    def and(
+      coderAssertion: CoderAssertion[T]
+    ): AssertionContext[T] = copy(lastAssertion = Some(coderAssertion(this)))
+  }
 
-  def roundtripKryo[T: ClassTag](
-    opts: PipelineOptions = DefaultPipelineOptions
-  ): CoderAssertion[T] =
-    new CoderAssertion[T] {
-      override def assert(value: T)(implicit c: Coder[T], eq: Equality[T]): Assertion = {
-        val kryoCoder = CoderMaterializer.beamWithDefault(Coder.kryo[T], o = opts)
-        checkRoundtripWithCoder(kryoCoder, value)
-      }
-    }
+  trait AssertionContextBase {
+    type ValType
+    val actualValue: Option[ValType]
+    val coder: Coder[ValType]
+    val lastAssertion: Option[Assertion]
+    val opts: PipelineOptions
+    lazy val beamCoder: BCoder[ValType] = CoderMaterializer.beamWithDefault(coder, opts)
+  }
 
-  def notFallback[T: ClassTag](opts: PipelineOptions = DefaultPipelineOptions): CoderAssertion[T] =
-    new CoderAssertion[T] {
-      override def assert(value: T)(implicit c: Coder[T], eq: Equality[T]): Assertion = {
-        c should !==(Coder.kryo[T])
-        val beamCoder = CoderMaterializer.beamWithDefault(c, o = opts)
-        checkRoundtripWithCoder[T](beamCoder, value)
-      }
-    }
+  def roundtrip[T: Equality](): CoderAssertion[T] = ctx =>
+    checkRoundtripWithCoder[T](ctx.beamCoder, ctx.actualValue.get)
 
-  def fallback[T: ClassTag](opts: PipelineOptions = DefaultPipelineOptions): CoderAssertion[T] =
-    new CoderAssertion[T] {
-      override def assert(value: T)(implicit c: Coder[T], eq: Equality[T]): Assertion = {
-        c should ===(Coder.kryo[T])
-        roundtripKryo(opts).assert(value)
-      }
-    }
+  def roundtripToBytes[T: Equality](expectedBytes: Array[Byte]): CoderAssertion[T] = ctx =>
+    checkRoundtripWithCoder[T](ctx.beamCoder, ctx.actualValue.get, expectedBytes)
 
-  def beConsistentWithEquals[T: ClassTag](
-    opts: PipelineOptions = DefaultPipelineOptions
-  ): CoderAssertion[T] =
-    new CoderAssertion[T] {
-      override def assert(value: T)(implicit c: Coder[T], eq: Equality[T]): Assertion = {
-        val beamCoder = CoderMaterializer.beamWithDefault(c, o = opts)
-        beamCoder.consistentWithEquals() shouldBe true
-      }
-    }
+  def haveCoderInstance(expectedCoder: Coder[_]): CoderAssertionBase = ctx =>
+    ctx.coder should ===(expectedCoder)
 
-  def beDeterministic[T: ClassTag](
-    opts: PipelineOptions = DefaultPipelineOptions
-  ): CoderAssertion[T] =
-    new CoderAssertion[T] {
-      override def assert(value: T)(implicit c: Coder[T], eq: Equality[T]): Assertion = {
-        val beamCoder = CoderMaterializer.beamWithDefault(c, o = opts)
-        noException should be thrownBy beamCoder.verifyDeterministic()
-      }
-    }
+  def notFallback[T: ClassTag: Equality](): CoderAssertion[T] = ctx => {
+    ctx.coder should !==(Coder.kryo[T])
+    checkRoundtripWithCoder(ctx.beamCoder, ctx.actualValue.get)
+  }
+
+  def fallback[T: ClassTag: Equality](): CoderAssertion[T] = ctx => {
+    ctx.coder should ===(Coder.kryo[T])
+    checkRoundtripWithCoder(ctx.beamCoder, ctx.actualValue.get)
+  }
+
+  def beConsistentWithEquals(): CoderAssertionBase = ctx =>
+    ctx.beamCoder.consistentWithEquals() shouldBe true
+
+  def beNotConsistentWithEquals(): CoderAssertionBase = ctx =>
+    ctx.beamCoder.consistentWithEquals() shouldBe false
+
+  def beDeterministic(): CoderAssertionBase = ctx =>
+    noException should be thrownBy ctx.beamCoder.verifyDeterministic()
+
+  def beNonDeterministic(): CoderAssertionBase = ctx =>
+    a[NonDeterministicException] should be thrownBy ctx.beamCoder.verifyDeterministic()
+
+  def beSerializable(): CoderAssertionBase = ctx =>
+    noException should be thrownBy SerializableUtils.ensureSerializable(ctx.beamCoder)
 
   def coderIsSerializable[A](implicit c: Coder[A]): Assertion =
-    coderIsSerializable(CoderMaterializer.beamWithDefault(c))
+    c.coderShould(beSerializable()).lastAssertion.get
 
-  private def coderIsSerializable[A](beamCoder: BCoder[A]): Assertion =
-    noException should be thrownBy SerializableUtils.ensureSerializable(beamCoder)
+  def beOfType[ExpectedCoder: ClassTag]: CoderAssertionBase = ctx =>
+    ctx.coder shouldBe a[ExpectedCoder]
 
-  private def checkRoundtripWithCoder[T](beamCoder: BCoder[T], value: T)(implicit
-    eq: Equality[T]
-  ): Assertion = {
-    val bytes = CoderUtils.encodeToByteArray(beamCoder, value)
-    val result = CoderUtils.decodeFromByteArray(beamCoder, bytes)
+  def materializeTo[ExpectedBeamCoder: ClassTag]: CoderAssertionBase =
+    ctx => {
+      ctx.beamCoder shouldBe a[MaterializedCoder[_]]
+      ctx.beamCoder.asInstanceOf[MaterializedCoder[_]].bcoder shouldBe a[ExpectedBeamCoder]
+    }
 
-    result should ===(value)
+  def materializeToTransformOf[ExpectedBeamCoder: ClassTag]: CoderAssertionBase =
+    ctx => {
+      ctx.beamCoder shouldBe a[MaterializedCoder[_]]
+      ctx.beamCoder.asInstanceOf[MaterializedCoder[_]].bcoder shouldBe a[TransformCoder[_, _]]
+      val innerCoder =
+        ctx.beamCoder.asInstanceOf[MaterializedCoder[_]].bcoder.asInstanceOf[TransformCoder[_, _]]
+      innerCoder.bcoder shouldBe a[ExpectedBeamCoder]
+    }
+
+  /*
+   * Checks that Beam's registerByteSizeObserver() and encode() are consistent
+   * */
+  def bytesCountTested[T <: Object: ClassTag](): CoderAssertion[T] =
+    ctx => {
+      val arr = Array(ctx.actualValue.get)
+      noException should be thrownBy CoderProperties.testByteCount(
+        ctx.beamCoder,
+        BCoder.Context.OUTER,
+        arr
+      )
+    }
+
+  /**
+   * Verifies that for the given coder and values, the structural values are equal if and only if
+   * the encoded bytes are equal. Verifies for Outer and Nested contexts
+   */
+  def structuralValueConsistentWithEquals(): CoderAssertionBase = ctx => {
+    noException should be thrownBy CoderProperties.structuralValueConsistentWithEquals(
+      ctx.beamCoder,
+      ctx.actualValue.get,
+      ctx.actualValue.get
+    )
   }
 
-  private def roundtripWithCoder[T](beamCoder: BCoder[T], value: T): T = {
-    val bytes = CoderUtils.encodeToByteArray(beamCoder, value)
-    CoderUtils.decodeFromByteArray(beamCoder, bytes)
+  /** Passes all checks on Beam coder */
+  def beFullyCompliant[T <: Object: ClassTag](): CoderAssertion[T] = ctx => {
+    structuralValueConsistentWithEquals()(ctx)
+    beSerializable()(ctx)
+    beConsistentWithEquals()(ctx)
+    bytesCountTested[T]().apply(ctx)
+    beDeterministic()(ctx)
+  }
+
+  def beFullyCompliantNonDeterministic[T <: Object: ClassTag](): CoderAssertion[T] =
+    ctx => {
+      structuralValueConsistentWithEquals()(ctx)
+      beSerializable()(ctx)
+      beConsistentWithEquals()(ctx)
+      bytesCountTested[T]().apply(ctx)
+      beNonDeterministic()(ctx)
+    }
+
+  def beFullyCompliantNotConsistentWithEquals[T <: Object: ClassTag](): CoderAssertion[T] =
+    ctx => {
+      structuralValueConsistentWithEquals()(ctx)
+      beSerializable()(ctx)
+      beNotConsistentWithEquals()(ctx)
+      bytesCountTested[T]().apply(ctx)
+      beDeterministic()(ctx)
+    }
+
+  private def checkRoundtripWithCoder[T: Equality](
+    beamCoder: BCoder[T],
+    actualValue: T,
+    expectedBytes: Array[Byte] = null
+  ): Assertion = {
+    val bytes = CoderUtils.encodeToByteArray(beamCoder, actualValue)
+    if (expectedBytes != null) {
+      bytes should ===(expectedBytes)
+    }
+    val result = CoderUtils.decodeFromByteArray(beamCoder, bytes)
+    result should ===(actualValue)
   }
 }
